backend/requirements.txt
fastapi
sqlalchemy
uvicorn
bcrypt
passlib
python-jose
pymysql
psycopg2